Moved CompressFileDialogTextValidator to it's own class...
Moved CompressFileDialogTextValidator to it's own class and fixed some
mistakes
