+++ app/src/main/java/com/amaze/filemanager/ui/dialogs/GeneralDialogCreation.java
@@ -15,7 +15,6 @@ import android.support.design.widget.TextInputEditText;
 import android.support.design.widget.TextInputLayout;
 import android.support.v7.widget.AppCompatButton;
 import android.support.v7.widget.AppCompatEditText;
-import android.text.Editable;
 import android.text.InputType;
 import android.text.SpannableString;
 import android.text.TextUtils;
@@ -51,11 +50,11 @@ import com.amaze.filemanager.filesystem.compressed.CompressedHelper;
 import com.amaze.filemanager.fragments.AppsListFragment;
 import com.amaze.filemanager.fragments.MainFragment;
 import com.amaze.filemanager.fragments.preference_fragments.PreferencesConstants;
+import com.amaze.filemanager.ui.views.CompressFileDialogTextValidator;
 import com.amaze.filemanager.utils.DataUtils;
 import com.amaze.filemanager.utils.FingerprintHandler;
 import com.amaze.filemanager.utils.OpenMode;
 import com.amaze.filemanager.utils.RootUtils;
-import com.amaze.filemanager.utils.SimpleTextWatcher;
 import com.amaze.filemanager.utils.Utils;
 import com.amaze.filemanager.utils.color.ColorUsage;
 import com.amaze.filemanager.utils.files.CryptUtil;
@@ -823,65 +822,6 @@ public class GeneralDialogCreation {
             b.getActionButton(DialogAction.NEGATIVE).setEnabled(false);
         }
         b.show();
-    }
-
-    static final Pattern ZIP_FILE_REGEX = Pattern.compile("[\\\\\\/:\\*\\?\"<>\\|\\x01-\\x1F\\x7F]", Pattern.CASE_INSENSITIVE);
-
-    static final class CompressFileDialogTextValidator extends SimpleTextWatcher implements View.OnFocusChangeListener
-    {
-        final WeakReference<Context> mContext;
-        final EditText mEditText;
-        final TextInputLayout mTextInputLayout;
-        final View mPositiveButton;
-
-        CompressFileDialogTextValidator(Context context, EditText editText, TextInputLayout textInputLayout, View positiveButton)
-        {
-            mContext = new WeakReference<Context>(context);
-            mEditText = editText;
-            mTextInputLayout = textInputLayout;
-            mPositiveButton = positiveButton;
-        }
-
-        private void doValidate()
-        {
-            Context context = mContext.get();
-            String value = mEditText.getText().toString();
-            //It's not easy to use regex to detect single/double dot while leaving valid values (filename.zip) behind...
-            //So we simply use equality to check them
-            boolean isValidFilename = (!ZIP_FILE_REGEX.matcher(value).find()) && !".".equals(value) && !"..".equals(value);
-
-            if(isValidFilename && value.length() > 0) {
-                mTextInputLayout.setError(null);
-                mPositiveButton.setEnabled(true);
-                if(!value.toLowerCase().endsWith(".zip"))
-                    mTextInputLayout.setHint(context.getResources().getString(R.string.compress_file_suggest_zip_extension));
-                else
-                    mTextInputLayout.setHint(context.getResources().getString(R.string.enterzipname));
-            } else {
-                mTextInputLayout.setHint(context.getResources().getString(R.string.enterzipname));
-
-                if(!isValidFilename)
-                    mTextInputLayout.setError(String.format(context.getText(R.string.invalid).toString(),
-                            context.getString(R.string.filename)));
-                else if(value.length() < 1)
-                    mTextInputLayout.setError(String.format(context.getText(R.string.cantbeempty).toString(),
-                            context.getString(R.string.filename)));
-
-                mPositiveButton.setEnabled(false);
-            }
-        }
-
-        @Override
-        public void afterTextChanged(Editable s) {
-            doValidate();
-        }
-
-        @Override
-        public void onFocusChange(View v, boolean hasFocus)
-        {
-            if(!hasFocus)
-                doValidate();
-        }
     }
 
     public static void showCompressDialog(final MainActivity m, final ArrayList<HybridFileParcelable> b, final String current) {
